requirements.txt
streamlit
Pillow
easyocr
numpy